+++ Twig/PagerfantaRuntime.php
@@ -2,11 +2,11 @@
 
 namespace BabDev\PagerfantaBundle\Twig;
 
-use BabDev\PagerfantaBundle\RouteGenerator\RouteGeneratorFactoryInterface;
-use BabDev\PagerfantaBundle\RouteGenerator\RouteGeneratorInterface;
 use Pagerfanta\Exception\OutOfRangeCurrentPageException;
 use Pagerfanta\Pagerfanta;
 use Pagerfanta\PagerfantaInterface;
+use Pagerfanta\RouteGenerator\RouteGeneratorFactoryInterface;
+use Pagerfanta\RouteGenerator\RouteGeneratorInterface;
 use Pagerfanta\View\ViewFactoryInterface;
 
 final class PagerfantaRuntime